requirements.txt
accelerate==1.3.0
annotated-types==0.7.0
attrs==25.1.0
beautifulsoup4==4.13.3
bioc==2.1
blinker==1.9.0
blis==1.2.0
boto3==1.36.19
botocore==1.36.19
breadability==0.1.20
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
conllu==4.5.3
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
Deprecated==1.2.18
docopt==0.6.2
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.17.0
flair==0.15.1
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.2.0
ftfy==6.3.1
gdown==5.2.0
gunicorn # Add this line
geojson==2.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
googlemaps==4.10.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
huggingface-hub==0.28.1
idna==3.10
intervaltree==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.6.0
mpld3==0.5.10
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pillow==11.1.0
pptree==3.1
praw==7.8.1
prawcore==2.4.0
preshed==3.0.9
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyowm==3.3.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch_revgrad==0.2.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.2
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
segtok==1.5.11
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
srsly==2.5.1
sumy==0.11.0
sympy==1.13.1
tabulate==0.9.0
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformer-smaller-training-vocab==0.4.0
transformers==4.48.3
typer==0.15.1
typing_extensions==4.12.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
Wikipedia-API==0.8.1
wrapt==1.17.2
yelpapi==2.5.1
